refactor ID datatype to be Integer

+++ backend/src/main/java/com/alancortez/project/repository/RecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
     Optional<Recipe> findByRecipeName(String recipeName);
 }
